+++ apps/public-docsite/src/components/Site/Site.tsx
@@ -96,7 +96,7 @@ export class Site<TPlatforms extends string = string> extends React.Component<
     let platform = 'default' as TPlatforms;
 
     // If current page doesn't have pages for the active platform, switch to its first platform.
-    if (Object.keys(navData.pagePlatforms).length > 0 && navData.activePages.length === 0) {
+    if (Object.keys(navData.pagePlatforms!).length > 0 && navData.activePages!.length === 0) {
       const firstPlatform = getPageFirstPlatform(getSiteArea(siteDefinition.pages), siteDefinition);
       const currentPage = getSiteArea(siteDefinition.pages);
       platform = firstPlatform;
@@ -142,7 +142,7 @@ export class Site<TPlatforms extends string = string> extends React.Component<
     const { siteDefinition } = this.props;
 
     // If current page doesn't have pages for the active platform, switch to its first platform.
-    if (Object.keys(pagePlatforms).length > 0 && activePages.length === 0) {
+    if (Object.keys(pagePlatforms!).length > 0 && activePages!.length === 0) {
       const firstPlatform = getPageFirstPlatform(getSiteArea(siteDefinition.pages), siteDefinition);
       this._onPlatformChanged(firstPlatform);
     }
@@ -347,22 +347,19 @@ export class Site<TPlatforms extends string = string> extends React.Component<
     return null;
   };
 
-  private _renderPlatformBar = (): JSX.Element | undefined => {
+  private _renderPlatformBar = (): JSX.Element | null => {
     const { siteDefinition } = this.props;
     const { platform, pagePlatforms, hasPlatformPicker } = this.state;
 
-    return (
-      hasPlatformPicker &&
-      Object.keys(pagePlatforms).length > 0 && (
-        <PlatformBar
-          activePlatform={platform}
-          onPlatformClick={this._onPlatformChanged}
-          pagePlatforms={pagePlatforms}
-          platforms={siteDefinition.platforms}
-          innerWidth={appMaximumWidthLg}
-        />
-      )
-    );
+    return hasPlatformPicker && Object.keys(pagePlatforms!).length > 0 ? (
+      <PlatformBar
+        activePlatform={platform}
+        onPlatformClick={this._onPlatformChanged}
+        pagePlatforms={pagePlatforms}
+        platforms={siteDefinition.platforms!}
+        innerWidth={appMaximumWidthLg}
+      />
+    ) : null;
   };
 
   /**
@@ -500,7 +497,7 @@ export class Site<TPlatforms extends string = string> extends React.Component<
     document.title = [
       siteDefinition.siteTitle,
       siteArea,
-      currPlatform && platforms[currPlatform]?.name,
+      currPlatform && platforms![currPlatform]?.name,
       activePageName !== siteArea && activePageName,
     ]
       .filter(Boolean)
@@ -531,7 +528,7 @@ export class Site<TPlatforms extends string = string> extends React.Component<
       this._jumpInterval = this._async.setInterval(() => {
         const el = document.getElementById(anchor);
         if (el || Date.now() - start > 1000) {
-          this._async.clearInterval(this._jumpInterval);
+          this._async.clearInterval(this._jumpInterval!);
           this._jumpInterval = undefined;
           if (el) {
             jumpToAnchor(anchor);