+++ apps/public-docsite/src/pages/Controls/DetailsListPage/DetailsListCompactPage.tsx
@@ -3,5 +3,5 @@ import { ControlsAreaPage, IControlsPageProps } from '../ControlsAreaPage';
 import { DetailsListCompactPageProps } from './DetailsListCompactPage.doc';
 
 export const DetailsListCompactPage: React.FunctionComponent<IControlsPageProps> = props => {
-  return <ControlsAreaPage {...props} {...DetailsListCompactPageProps[props.platform]} />;
+  return <ControlsAreaPage {...props} {...DetailsListCompactPageProps[props.platform!]} />;
 };