+++ apps/public-docsite/src/pages/Controls/DetailsListPage/DetailsListCustomRowsPage.tsx
@@ -3,5 +3,5 @@ import { ControlsAreaPage, IControlsPageProps } from '../ControlsAreaPage';
 import { DetailsListCustomRowsPageProps } from './DetailsListCustomRowsPage.doc';
 
 export const DetailsListCustomRowsPage: React.FunctionComponent<IControlsPageProps> = props => {
-  return <ControlsAreaPage {...props} {...DetailsListCustomRowsPageProps[props.platform]} />;
+  return <ControlsAreaPage {...props} {...DetailsListCustomRowsPageProps[props.platform!]} />;
 };