+++ apps/public-docsite/src/pages/Controls/SelectionPage/SelectionPage.tsx
@@ -3,5 +3,5 @@ import { ControlsAreaPage, IControlsPageProps } from '../ControlsAreaPage';
 import { SelectionPageProps } from './SelectionPage.doc';
 
 export const SelectionPage: React.FunctionComponent<IControlsPageProps> = props => {
-  return <ControlsAreaPage {...props} {...SelectionPageProps[props.platform]} />;
+  return <ControlsAreaPage {...props} {...SelectionPageProps[props.platform!]} />;
 };