+++ apps/public-docsite/src/pages/Styles/LocalizationPage/LocalizationPage.tsx
@@ -17,13 +17,13 @@ export const LocalizationPage: React.FunctionComponent<IStylesPageProps> = props
   return (
     <StylesAreaPage
       {...props}
-      {...LocalizationPageProps[platform]}
-      otherSections={_otherSections(platform) as IPageSectionProps[]}
+      {...LocalizationPageProps[platform!]}
+      otherSections={_otherSections(platform!) as IPageSectionProps[]}
     />
   );
 };
 
-function _otherSections(platform: Platforms): IPageSectionProps<Platforms>[] {
+function _otherSections(platform: Platforms): IPageSectionProps<Platforms>[] | undefined {
   switch (platform) {
     case 'web':
       return [