+++ packages/react-components/react-table/src/components/TableHeaderCell/useTableHeaderCellStyles.ts
@@ -63,8 +63,8 @@ const useStyles = makeStyles({
     flexGrow: 1,
     height: '100%',
     alignItems: 'center',
-    ...shorthands.gap(tokens.spacingHorizontalS),
-    minHeight: '44px',
+    ...shorthands.gap(tokens.spacingHorizontalXS),
+    minHeight: '32px',
     ...shorthands.flex(1, 1, '0px'),
   },
   sortable: {
@@ -74,6 +74,7 @@ const useStyles = makeStyles({
   sortIcon: {
     display: 'flex',
     alignItems: 'center',
+    paddingTop: tokens.spacingVerticalXXS,
   },
 });