+++ packages/react-components/react-table/stories/DataGrid/index.stories.tsx
@@ -1,12 +1,40 @@
-import { DataGrid } from '../../src/components/DataGrid/DataGrid';
+import {
+  DataGrid,
+  DataGridHeader,
+  DataGridHeaderCell,
+  DataGridBody,
+  DataGridRow,
+  DataGridCell,
+  DataGridSelectionCell,
+} from '@fluentui/react-components/unstable';
+import descriptionMd from './DataGridDescription.md';
 
 export { Default } from './Default.stories';
+export { KeyboardNavigation } from './KeyboardNavigation.stories';
+export { Sort } from './Sort.stories';
+export { SortControlled } from './SortControlled.stories';
+export { MultipleSelect } from './MultipleSelect.stories';
+export { MultipleSelectControlled } from './MultipleSelectControlled.stories';
+export { SingleSelect } from './SingleSelect.stories';
+export { SingleSelectControlled } from './SingleSelectControlled.stories';
+export { SubtleSelection } from './SubtleSelection.stories';
+export { SelectionAppearance } from './SelectionAppearance.stories';
 export default {
   title: 'Preview Components/DataGrid',
   component: DataGrid,
+  subcomponents: {
+    DataGridHeader,
+    DataGridHeaderCell,
+    DataGridBody,
+    DataGridRow,
+    DataGridCell,
+    DataGridSelectionCell,
+  },
   parameters: {
     docs: {
-      description: {},
+      description: {
+        component: [descriptionMd].join('\n'),
+      },
     },
   },
 };