+++ packages/react-components/react-tree/src/components/TreeLeaf/TreeLeaf.types.ts
@@ -1,9 +1,0 @@
-import type { ComponentProps, ComponentState, Slot } from '@fluentui/react-utilities';
-
-export type TreeLeafSlots = {
-  root: Slot<'div'>;
-};
-
-export type TreeLeafProps = ComponentProps<TreeLeafSlots> & {};
-
-export type TreeLeafState = ComponentState<TreeLeafSlots>;