+++ scripts/tasks/copy.ts
@@ -44,14 +44,15 @@ export function copyCompiled() {
   const packageDir = process.cwd();
 
   if (!(isUsingTsSolutionConfigs && tsConfig)) {
-    throw new Error(`this task compliant only with packages that use TS solution config files.`);
+    logger.warn(`copy-compiled: works only with packages that use TS solution config files. Skipping...`);
+    return;
   }
 
   // TODO: remove after all v9 is migrated to new build and .d.ts API stripping
   const hasNewCompilationSetup = (tsConfig.compilerOptions.outDir as string).includes('dist/out-tsc');
 
   if (!hasNewCompilationSetup) {
-    logger.info('copy-compiled: noop ');
+    logger.info('copy-compiled: noop');
 
     return;
   }