+++ packages/react-components/react-card/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log - @fluentui/react-card
 
-This log was last generated on Fri, 11 Nov 2022 14:57:56 GMT and should not be manually modified.
+This log was last generated on Thu, 17 Nov 2022 23:05:35 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.0.0-beta.33](https://github.com/microsoft/fluentui/tree/@fluentui/react-card_v9.0.0-beta.33)
+
+Thu, 17 Nov 2022 23:05:35 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-card_v9.0.0-beta.32..@fluentui/react-card_v9.0.0-beta.33)
+
+### Changes
+
+- Bump @fluentui/react-tabster to v9.3.1 ([PR #25683](https://github.com/microsoft/fluentui/pull/25683) by beachball)
+- Bump @fluentui/react-utilities to v9.2.2 ([PR #25683](https://github.com/microsoft/fluentui/pull/25683) by beachball)
+- Bump @fluentui/react-button to v9.1.9 ([PR #25683](https://github.com/microsoft/fluentui/pull/25683) by beachball)
 
 ## [9.0.0-beta.32](https://github.com/microsoft/fluentui/tree/@fluentui/react-card_v9.0.0-beta.32)