+++ packages/react-components/react-link/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui/react-link
 
-This log was last generated on Fri, 11 Nov 2022 14:58:07 GMT and should not be manually modified.
+This log was last generated on Thu, 17 Nov 2022 23:05:41 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.0.12](https://github.com/microsoft/fluentui/tree/@fluentui/react-link_v9.0.12)
+
+Thu, 17 Nov 2022 23:05:41 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-link_v9.0.11..@fluentui/react-link_v9.0.12)
+
+### Patches
+
+- Bump @fluentui/react-tabster to v9.3.1 ([PR #25683](https://github.com/microsoft/fluentui/pull/25683) by beachball)
+- Bump @fluentui/react-utilities to v9.2.2 ([PR #25683](https://github.com/microsoft/fluentui/pull/25683) by beachball)
 
 ## [9.0.11](https://github.com/microsoft/fluentui/tree/@fluentui/react-link_v9.0.11)