+++ packages/react-components/react-overflow/CHANGELOG.md
@@ -1,8 +1,20 @@
 # Change Log - @fluentui/react-overflow
 
-This log was last generated on Fri, 11 Nov 2022 14:58:08 GMT and should not be manually modified.
+This log was last generated on Thu, 17 Nov 2022 23:05:42 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.0.0-rc.1](https://github.com/microsoft/fluentui/tree/@fluentui/react-overflow_v9.0.0-rc.1)
+
+Thu, 17 Nov 2022 23:05:42 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-overflow_v9.0.0-beta.14..@fluentui/react-overflow_v9.0.0-rc.1)
+
+### Changes
+
+- feat: Bump to RC ([PR #25659](https://github.com/microsoft/fluentui/pull/25659) by lingfangao@hotmail.com)
+- Bump @fluentui/priority-overflow to v9.0.0-rc.1 ([PR #25683](https://github.com/microsoft/fluentui/pull/25683) by beachball)
+- Bump @fluentui/react-context-selector to v9.1.2 ([PR #25683](https://github.com/microsoft/fluentui/pull/25683) by beachball)
+- Bump @fluentui/react-utilities to v9.2.2 ([PR #25683](https://github.com/microsoft/fluentui/pull/25683) by beachball)
 
 ## [9.0.0-beta.14](https://github.com/microsoft/fluentui/tree/@fluentui/react-overflow_v9.0.0-beta.14)