+++ packages/react-components/react-portal-compat/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui/react-portal-compat
 
-This log was last generated on Fri, 11 Nov 2022 14:57:31 GMT and should not be manually modified.
+This log was last generated on Thu, 17 Nov 2022 23:05:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.0.19](https://github.com/microsoft/fluentui/tree/@fluentui/react-portal-compat_v9.0.19)
+
+Thu, 17 Nov 2022 23:05:43 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-portal-compat_v9.0.18..@fluentui/react-portal-compat_v9.0.19)
+
+### Patches
+
+- Bump @fluentui/react-tabster to v9.3.1 ([PR #25683](https://github.com/microsoft/fluentui/pull/25683) by beachball)
+- Bump @fluentui/react-components to v9.7.1 ([PR #25683](https://github.com/microsoft/fluentui/pull/25683) by beachball)
 
 ## [9.0.18](https://github.com/microsoft/fluentui/tree/@fluentui/react-portal-compat_v9.0.18)