+++ packages/react-migration-v8-v9/CHANGELOG.md
@@ -1,8 +1,62 @@
 # Change Log - @fluentui/react-migration-v8-v9
 
-This log was last generated on Tue, 15 Nov 2022 07:44:59 GMT and should not be manually modified.
+This log was last generated on Wed, 30 Nov 2022 07:46:31 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.0.16](https://github.com/microsoft/fluentui/tree/@fluentui/react-migration-v8-v9_v1.0.16)
+
+Wed, 30 Nov 2022 07:46:31 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-migration-v8-v9_v1.0.15..@fluentui/react-migration-v8-v9_v1.0.16)
+
+### Patches
+
+- Bump @fluentui/react to v8.103.3 ([PR #25836](https://github.com/microsoft/fluentui/pull/25836) by beachball)
+
+## [1.0.15](https://github.com/microsoft/fluentui/tree/@fluentui/react-migration-v8-v9_v1.0.15)
+
+Thu, 24 Nov 2022 07:53:48 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-migration-v8-v9_v1.0.14..@fluentui/react-migration-v8-v9_v1.0.15)
+
+### Patches
+
+- Bump @fluentui/react to v8.103.2 ([PR #25770](https://github.com/microsoft/fluentui/pull/25770) by beachball)
+
+## [1.0.14](https://github.com/microsoft/fluentui/tree/@fluentui/react-migration-v8-v9_v1.0.14)
+
+Wed, 23 Nov 2022 07:54:37 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-migration-v8-v9_v1.0.13..@fluentui/react-migration-v8-v9_v1.0.14)
+
+### Patches
+
+- Bump @fluentui/react to v8.103.1 ([PR #25726](https://github.com/microsoft/fluentui/pull/25726) by beachball)
+
+## [1.0.13](https://github.com/microsoft/fluentui/tree/@fluentui/react-migration-v8-v9_v1.0.13)
+
+Tue, 22 Nov 2022 07:47:00 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-migration-v8-v9_v1.0.12..@fluentui/react-migration-v8-v9_v1.0.13)
+
+### Patches
+
+- Bump @fluentui/react to v8.103.0 ([PR #25499](https://github.com/microsoft/fluentui/pull/25499) by beachball)
+
+## [1.0.12](https://github.com/microsoft/fluentui/tree/@fluentui/react-migration-v8-v9_v1.0.12)
+
+Thu, 17 Nov 2022 07:51:44 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-migration-v8-v9_v1.0.11..@fluentui/react-migration-v8-v9_v1.0.12)
+
+### Patches
+
+- Bump @fluentui/react to v8.102.0 ([PR #25638](https://github.com/microsoft/fluentui/pull/25638) by beachball)
+
+## [1.0.11](https://github.com/microsoft/fluentui/tree/@fluentui/react-migration-v8-v9_v1.0.11)
+
+Wed, 16 Nov 2022 07:47:29 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-migration-v8-v9_v1.0.10..@fluentui/react-migration-v8-v9_v1.0.11)
+
+### Patches
+
+- Bump @fluentui/react to v8.101.2 ([PR #25626](https://github.com/microsoft/fluentui/pull/25626) by beachball)
 
 ## [1.0.10](https://github.com/microsoft/fluentui/tree/@fluentui/react-migration-v8-v9_v1.0.10)